Store gridded latlons in float32 (previously float64)

+++ training/src/anemoi/training/data/data_reader.py
@@ -271,12 +271,12 @@ class GriddedDataReader(BaseAnemoiReader, ABC):
         Backed by ``self.data.latitudes`` (which is stored in degrees by
         ``anemoi.datasets``); converted once and cached.
         """
-        return np.deg2rad(np.asarray(self.data.latitudes))
+        return np.deg2rad(np.asarray(self.data.latitudes, dtype=np.float32))
 
     @cached_property
     def longitudes(self) -> np.ndarray:
         """Return per-grid-point longitudes in **radians**."""
-        return np.deg2rad(np.asarray(self.data.longitudes))
+        return np.deg2rad(np.asarray(self.data.longitudes, dtype=np.float32))
 
     @property
     def has_trajectories(self) -> bool:
